+++ docker/mongodb-enterprise-tests/tests/vaultintegration/mongodb_deployment_vault.py
@@ -183,6 +183,7 @@ def mongodb_user(namespace: str) -> MongoDBUser:
     }
 
     resource["spec"]["mongodbResourceRef"]["name"] = MDB_RESOURCE
+    resource["spec"]["mongodbResourceRef"]["namespace"] = namespace
     return resource.create()